Refactoring project structure add Dockerfile add docker command's to makefile
