+++ app/tests/test_main.py
@@ -1,6 +1,6 @@
 from fastapi.testclient import TestClient
 
-from .main import app
+from ..src.main import app
 
 
 client = TestClient(app)
